zwp_text_input_v3: Fix text deletion
This change tries to fix text deletion in absence of set_surrounding_text support.

Deleting text using byte ranges is only possible when set_surrounding_text is available and properly supported. However, it's not mandated, and mandating it would increase the difficulty threshold of implementing the protocol.

The solution to that is to redefine text deletion requests to use a unit that can be requested without knowing the contents. The ideal unit would be some a "character", however Unicode doesn't offer a good definition of one. To avoid this, the request has been overloaded to let the client decide what a "character" is when surrounding text is not supported, and let the compositor do that when it is, by deleting using Unicode codepoint count.

To accommodate for giving the compositor such responsibility, a minimal requirement is placed on the amount of text in set_surrounding_text.