ext-blur: New protocol
This patch introduces a protocol that provides a way to specify the area
of a surface that needs to have its background blurred by the compositor.

An intended use case is to improve visuals of translucent surfaces such
as panels or docks by blurring background behind them.

This protocol is based on the org_kde_kwin_blur protocol, which has been
in use since 2015.

Signed-off-by: Vlad Zahorodnii <vlad.zahorodnii@kde.org>