DBus annotation protocol
The protocol allows clients or surfaces to be associated
with DBus objects and interfaces.

This formalizes the ad-hoc extensions made to X
and similar compositor-specific Wayland protocols
for assorted tasks that boil down to associating
DBus objects with windows/surfaces.

In user-facing terms, this allows Unity-style application
menus to be implemented in a cross-compositor manner
on Wayland, and provides a standard mechanism for applications
such as Firefox and Google Chrome to program against.

Signed-off-by: Janet Blackquill <uhhadd@gmail.com>