Add a configure event to zwp_input_panel_v1
When displaying an input panel, the compositor may sometimes want to
adjust the placement and size of the input panel, for example to account
for a task bar or other shell surface that the input panel may overlap.
While placement is already done by the compositor, currently there is no
way to communicate a desired size to the input panel client, so sizing
is not possible.

This adds a configure event to `zwp_input_panel_v1`, allowing the
compositor to request a new panel size from the input panel client. A
corresdponding ack_configure request is also added so the client can let
the compositor know it processed the configure.

Signed-off-by: Arjen Hiemstra <ahiemstra@heimr.nl>