wp-fractional-scale-v2: fractional scaling support
Fractional scaling is a way of scaling user interfaces that allows for non-integer
scaling factors. In order to support multiple displays with different fractional
scaling factors properly and prevent size mismatches in all relevant situations,
multiple coordinate spaces need to be used. This commit adds a protocol
that adds the ability for compositor and client to change the coordinate
space of their communication for that purpose.

Signed-off-by: Xaver Hugl <xaver.hugl@gmail.com>