color: re-design ICC image description factory
Turn wp_image_description_creator_icc_v1 into a global interface, so
that compositors can choose whether to advertise it or not. The name is
changed for consistency:
- global interfaces are factories
- one-shot helper objects are creators

set_icc request is turned into simply "create" that creates the image
description object. There is no room for any additional parameters,
because I believe we need to let the ICC handle those. We should not be
modifying their intents by adding e.g. viewing environment variables
here. What the ICC does not define, we do not add.

ICC data offset and length are added, so that ICC profiles embedded in
image files are easier to just pass through. Offset is only 32 bits,
though.

Some protocol errors that are clearly client bugs are defined, but all
other interpretation errors are left for the graceful failure event to
handle.

Add a definition of when a compositor must be finished with reading the
fd, to let clients know when it's safe to e.g. modify the file again.

The creation of the wp_image_description_v1 object requires the client
to set the interface and version explicitly. This is to allow extending
wp_image_description_v1 interface while allowing multiple independent
factories for it: the interface version must be given instead of
inherited.

Details of the error event are to be clarified later.

Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.com>