color: rework how wp_image_description_creator_params_v1 is created
Since the params object holds mutable state, it would be unorthodox to
create that straight from a global via wl_registry.bind, even though we
could just fine. To make things more familiar, define a separate global
factory interface for params objects.

Define the fatal protocol errors for this interface that are definitely
client bugs rather than unsupported image descriptions. Unsupported
image description errors are punted to a future image description
graceful error event that the ICC factory is going to need as well.

The error codes for duplicate setting of properties will be added later
as fatal protocol errors. More protocol errors for universally
unsupported image descriptions might be added later too.

The created and error events are removed, because the error event will
be in wp_image_description_v1 interface, and we cannot use the created
event because we need to pass the interface version explicitly to avoid
inheriting it. Unqualified new_id argument in an event has never been
tried as I know of, but it works in requests and causes less surprises.

The create request could be a second destructor for this interface, but
it's not a common pattern to have destructor with arguments, even less
with a new_id, so I chose to not try my luck there. Semantically it
would be correct. So, we do what zwp_linux_buffer_params_v1 does and
just say you need to destroy the object.

Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.com>