color: update wp_color_management_output_v1.get_image_description
Update how the object must be created so it has an explicit interface
version. This might not be strictly necessary here because the inherited
version would be likely correct, but it's done for consistency.

Take into account that if wp_image_description_v1 interface does need
extending, old versions of it may not be sufficient for describing
outputs.

Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.com>