color: drop cicp_color_primaries_id enum
Since we are going to rely on H.273 CICP, we do not need to have an
enumeration defined for these, we will simply use the CICP values.

It's also problematic to create names for the CICP values, because each
code point matches several standards so they don't have canonical names.

Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.com>