color: reword icc_file event
Allow the fd to be read(2) as well. Mention that the compositor chooses
the profile details. It is possible that a compositor was not configured
with an output ICC profile, so it cannot sent one in that case.

Add a note about offset, which is different from
wp_image_description_factory_icc_v1: a compositor cannot risk some
client reading data before the given offset leading to an information
leak, so for non-zero offset a compositor would need to clear that
memory, so cannot use any bigger file as the base anyway.

Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.com>