color: clarify wp_image_description_v1.destroy
Move it up so it's the first defined request. Destruction is such a
fundamental part of any interface.

Make it more explicit on what destructions means. I imagine compositors
to use reference-counted image description implementation objects that
are independent of protocol objects. That is already what 'identity'
event implies. It would be more hassle to make sure that
set_image_description, destroy, commit would actually special-case the
destroy, and we would also need to define what that means.

Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.com>