staging/xdg-activation: add a request for choosing windows to activate
Some clients allow having multiple documents in a single window, while also
allowing multiple windows at the same time. When such a client is asked
to open a new document by another application, it needs to choose a window
to open it in, but it doesn't have enough contextual information to do that
well in all cases. It can for example not know if a window is currently
on an inactive virtual desktop.

The added request solves this problem by having the client provide an
activation token and a list of windows, so that the compositor can choose
the best window for opening a new document for the application. It may also
not choose any window, in which case the application should create a new
window.

Signed-off-by: Xaver Hugl <xaver.hugl@gmail.com>