color: add supported_feature event
This makes the support for a few requests to be explicitly advertised
and unavailable otherwise. This allows compositors to support only
enumerated transfer functions and primaries, instead of requiring them
to support the arbitrary flavors. This may be useful for
hardware-specific compositors that cannot use programmable shading.

set_mastering_display_primaries is a bit more special. The support for
it is split into two features: basic support, and extended color volume
support. Extended color volume support requires handling color values
outside of the range [0.0, 1.0] which requires special care from
compositors. Silently clipping out-of-range values may or may not be
acceptable, so clients should be aware whether it works. If clipping is
acceptable, they can try an image description without extended color
volume.

Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.com>