color: simplify factory design
Supporting other factories in independent protocol extensions was a nice
idea, but complicates the design for questionable benefit. The existing
design was only half of it, too, because wp_image_description_v1
interface had events related to all factories instead of the factories
having their own events.

Therefore go back to a simple monolithic design, but keep the different
factory requests still optional and discoverable.

This gets rid of the new_id interface hack.

There is also a little bit of unifying the wording, and couple typo
fixes.

The ICC factory is turned into a creator object, because we are not sure
that there cannot be other properties that would need to be set.
When a compositor can read the file is clarified.

Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.com>