unstable/text-input: Add preedit text hints
There seems to be some demand from input methods to be able
to style different parts of the preedit string differently.
Since the reasons to do that are many, and vary between IMs
and languages, there may not ever be a closed semantic list
of hints.

But even though some IMs would much want to be able to specify
colors, they lack the context to know what choice of colors
provides enough contrast and legibility embedded on foreign
UIs (e.g. accounting things like themes, accessibility, ...).

This MR meets at a middle ground, and provides a set of wildly
generic semantic values, so that clients may theme these
specifically, and the IM is able to provide some hints rather
than none.

Control over underlining is given to the IMs, since that is
less controversial.

Signed-off-by: Carlos Garnacho <carlosg@gnome.org>