staging: add toplevel tag protocol
The window id protocol allows clients to set a tag for toplevels, which
the compositor can use to identify them even after the application
has been restarted. This persistent identification can be used by the compositor
to restore properties like position, size, "always on top", and it can also
be used for allowing users to create rules that change compositor behavior
for specific windows.

Signed-off-by: Xaver Hugl <xaver.hugl@kde.org>