Migrate selected protocols from unstable/ to staging/
Many protocols in "unstable/" are not really unstable by any means, and
are not expected to change in any backward incompatible way. The fact
that they are still in unstable/ is awkward from a communication
perspective, since it incorrectly communicates that they are actually
expected to change.

Migrating protocols that are used in production that are not expecting
to change to staging/ solves this issue.

This commit migrates the following protocols from unstable/ to staging/,
updating the disclaimer accordingly.

 * idle-inhibit/idle-inhibit-v1.xml
 * input-timestamps/input-timestamps-v1.xml
 * keyboard-shortcuts-inhibit/keyboard-shortcuts-inhibit-v1.xml
 * pointer-constraints/pointer-constraints-v1.xml
 * pointer-gestures/pointer-gestures-v1.xml
 * primary-selection/primary-selection-v1.xml
 * relative-pointer/relative-pointer-v1.xml
 * text-input/text-input-v3.xml
 * xdg-decoration/xdg-decoration-v1.xml
 * xdg-foreign/xdg-foreign-v2.xml

Some protocols were explicitly left out:

 * fullscreen-shell/fullscreen-shell-unstable-v1.xml
 * input-method/input-method-unstable-v1.xml
 * tablet/tablet-unstable-v1.xml
 * text-input/text-input-unstable-v1.xml
 * xdg-foreign/xdg-foreign-unstable-v1.xml
 * linux-dmabuf/linux-dmabuf-unstable-v1.xml
 * linux-explicit-synchronization/linux-explicit-synchronization-unstable-v1.xml

The reason are either that the protocol is its current form is abandoned
(fullscreen shell), has been replaced with a new major version
(text-input v1, xdg-foreign v1, tablet v1), is being migrated
independently (linux-dmabuf), has a backward incompatible replacement
in active development (linux-explicit-synchronization), or have more
complicated story (input-method).

I also left out xwayland-keyboard-grab, as it's unclear whether this is
a protocol that should be fully owned by Xwayland or not.

Some protocols were missing toplevel description, so added those and
moved the disclaimer there.

Signed-off-by: Jonas Ådahl <jadahl@gmail.com>