color: Define out of bounds pixel values
When values are out of the defined transfer function range they cannot
be processed by the definition. State this, and recommend a workaround
to avoid producing an invalid composite image.

Clamping is not required, because the requirement might not be
fulfillable by display hardware via KMS. We don't want to lose KMS
off-loading opportunities by a mere possibility of badly formed data.

Signed-off-by: Sebastian Wick <sebastian.wick@redhat.com>
[Pekka: made clamping optional]
Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.com>