xdg-shell: Enforce geometry for tiled xdg_toplevel
Tiling indicates that the respective toplevel edge is in contact with
some form of boundary, which could for example be another toplevel or
the edge of a monitor. If a toplevel has boundaries on two opposing
edges, the toplevel should strictly adhere to the relevant dimension
from xdg_toplevel.configure to fill the space available between these
boundaries without exceeding it.