color: inert get_image_description is error
Since this is a nonsensical action the client is doing, make it a
protocol error: the client has first destroyed the wl_output proxy,
then proceeds to ask the current image description of that wl_output.

This is slightly against the inert object semantics, but the alternative
would be to gracefully fail the image description in a situation that is
guaranteed to fail. Better to not allow it in the first place.

The corner case was found during the Weston implementation.

Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.com>