color: inert get_preferred is error
Make it a protocol error to first destroy the associated wl_surface, and
then ask for the preferred image description on that surface.

This is slightly against the inert object semantics, but the alternative
would be to gracefully fail the image description in a situation that is
guaranteed to fail. Better to not allow it in the first place.

Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.com>