xdg-shell: relax maximized size constraints
In practice, no compositor is disconnecting clients with a protocol
error when they don't obey the size dictated by the compositor in
maximized state. It's very unlikely that any compositor would start
doing so, as this would kill many existing clients. Moreover, some
clients have a maximum size (e.g. dialogs), and the compositor is
still allowed to mark them as maximized, in which case it's not clear
how the client should react if the configured size exceeds the
maximum size.

Update the protocol to reflect reality. Since we still want clients
to obey the size in the general case, keep the "must".

Signed-off-by: Simon Ser <contact@emersion.fr>