tablet: Add scroll events to the tablet interface
A common-enough interaction with the tablet is to use the tablet for
scrolling - either "pan-scrolling" (one specific pen button down, moving
the pen scroll) or ring-scrolling/strip-scrolling.

Neither are supported in the protocol right now and we completely rely
on the client to implement this interaction - arguably this should be
supported more generically in the compositor to have the same
interactions work the same way.