cursor-shape: Add the 'all_resize' cursor shape
This is a cursor shape that can be used to indicate that a window
or other object can be moved in any direction. A typical image
for this cursor is a 4-headed arrow with horizontal and vertical
arrows pointing in all 4 directions. This image is also often
used for the 'move' cursor, but that is problematic because it
is not visually compatible with the other drag-and-drop cursors,
so we introduce a new name to separate the two use cases.

Fixes: #204

Signed-off-by: Matthias Clasen <mclasen@redhat.com>