xdg-shell: document resize with "none" edges
Current compositor behavior here is not terribly consistent:

Both Weston and Mutter ignore resize requests with "none" edges.

wlroots currently kills the client with a protocol error but allowed
compositors using the wlroots library to decide what to do in the past
before this behavior was accidentally changed. Sending a protocol here
is a wlroots bug in my opinion since the none enum variant exists and
the invalid_resize_edge protocol error is only specified to be sent for
values that do not match an enum variant.

KWin seems to treat a resize with "none" edges the same as a move
request, which is quite strange. I would consider this a bug as well.

This patch recommends that compositors ignore resize requests with
"none" edges in the interest of increasing client portability.

Signed-off-by: Isaac Freund <mail@isaacfreund.com>