xdg-toplevel-drag-v1: add set_offset request
This request allows the client to change the offset hint without
unmapping the toplevel. This is useful when the geometry of the toplevel
changes while it is being dragged.

To avoid jank, the hint change is integrated with the commit sequence
that changes the geometry.

Additionally, the client must now perform a commit after attaching a
toplevel to ensure that the offset in the attach request is also in sync
with the geometry of the window.

Signed-off-by: Julian Orth <ju.orth@gmail.com>