commit-timing/fifo: Add more pragmatic synchronized subsurface semantics
We need to do things like pausing a video player or resizing an application
window that contains subsurfaces. If queues of future updates exist, we
don't want to have to wait for them to complete to perform these
operations.

At these times it would be useful to have a way to stop the application
of future updates immediately.

Since we can't un-make a content update, we also need semantics to
fast-foward through them, so we can provide a new current update and
appear visibly to have "skipped" everything in between.

Signed-off-by: Derek Foreman <derek.foreman@collabora.com>